backend/requirements.txt
fastapi
uvicorn[standard]
face-recognition
opencv-python-headless
numpy
supabase
python-dotenv
fpdf
python-multipart
